refactor(init/algebra): use more biconditional algebra lemmas
This is an attempt to clean up the many one-way algebra theorems and encourage a more efficient collection of theorems that are all biconditional. These have the advantage that they can be used as simp rules, they chain together with transitivity, and we can access the one-way rules easily with .1 or .2.